fix(build/wrap.go): make it work on Apple Silicon (#3)
On Apple Silicon systems, Homebrew installs under /opt. We need to tell tor's configure about that.

While there, make the lockfile more pleasant to use.

While there, make sure we follow the main-0.4.7 branch of tor.

This work is under the https://github.com/ooni/probe/issues/2365 umbrella. While I managed to get rid of go-libtor for Android, doing the same for iOS seems a bit cumbersome, and I already spent lots of time on this task. Therefore, I am going to update go-libtor such that we can use it on iOS systems. This diff implements the first part of this N-th side task, that is, making sure I am a bit more comfortable using this library and I can update it confidently.